Move image MIME type constants to Tools\Images\Type
