+++ includes/avatar-privacy/default-icons/class-custom-icon-provider.php
@@ -29,11 +29,9 @@ namespace Avatar_Privacy\Default_Icons;
 use Avatar_Privacy\Core;
 use Avatar_Privacy\Settings;
 
-use Avatar_Privacy\Components\Image_Proxy;
-
 use Avatar_Privacy\Data_Storage\Filesystem_Cache;
 
-use Avatar_Privacy\Tools\Images as Image_Tools;
+use Avatar_Privacy\Tools\Images;
 
 use Avatar_Privacy\Upload_Handlers\Custom_Default_Icon_Upload_Handler as Upload;
 
@@ -101,15 +99,15 @@ class Custom_Icon_Provider extends Abstract_Icon_Provider {
 
 		// We need the current site ID.
 		$site_id   = \get_current_blog_id();
-		$extension = Image_Proxy::FILE_EXTENSION[ $icon['type'] ];
+		$extension = Images\Type::FILE_EXTENSION[ $icon['type'] ];
 		$identity  = $this->core->get_hash( "custom-default-{$site_id}" );
 		$filename  = "custom/{$site_id}/{$identity}-{$size}.{$extension}";
 
 		// Only generate a new icon if necessary.
 		if ( ! \file_exists( "{$this->file_cache->get_base_dir()}{$filename}" ) ) {
 
-			$data = Image_Tools\Editor::get_resized_image_data(
-				Image_Tools\Editor::get_image_editor( $icon['file'] ), $size, $size, true, $icon['type']
+			$data = Images\Editor::get_resized_image_data(
+				Images\Editor::get_image_editor( $icon['file'] ), $size, $size, true, $icon['type']
 			);
 			if ( empty( $data ) ) {
 				// Something went wrong..
