+++ includes/avatar-privacy/tools/images/class-type.php
@@ -1,0 +1,56 @@
+<?php
+/**
+ * This file is part of Avatar Privacy.
+ *
+ * Copyright 2018 Peter Putzer.
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU General Public License
+ * as published by the Free Software Foundation; either version 2
+ * of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+ *
+ *  ***
+ *
+ * @package mundschenk-at/avatar-privacy
+ * @license http://www.gnu.org/licenses/gpl-2.0.html
+ */
+
+namespace Avatar_Privacy\Tools\Images;
+
+/**
+ * An abstract class for MIME type and extension handling.
+ *
+ * @since 1.2.0
+ *
+ * @author Peter Putzer <github@mundschenk.at>
+ */
+abstract class Type {
+	const JPEG_IMAGE = 'image/jpeg';
+	const PNG_IMAGE  = 'image/png';
+	const SVG_IMAGE  = 'image/svg+xml';
+
+	const JPEG_EXTENSION = 'jpg';
+	const PNG_EXTENSION  = 'png';
+	const SVG_EXTENSION  = 'svg';
+
+	const CONTENT_TYPE = [
+		self::JPEG_EXTENSION => self::JPEG_IMAGE,
+		self::PNG_EXTENSION  => self::PNG_IMAGE,
+		self::SVG_EXTENSION  => self::SVG_IMAGE,
+	];
+
+	const FILE_EXTENSION = [
+		self::JPEG_IMAGE => self::JPEG_EXTENSION,
+		self::PNG_IMAGE  => self::PNG_EXTENSION,
+		self::SVG_IMAGE  => self::SVG_EXTENSION,
+	];
+}
